fixed 로그인 후 페이지 이동하지 않고 싱글페이지에서 작동하도록 개선 - 기타 소스 코드 리팩토링
