+++ src/RoutesSet.js
@@ -1,14 +1,11 @@
 import { Route, Routes } from "react-router-dom"
-
-import LogIn from './components/LogIn/LogIn'
-import FormationBoard from './pages/FormationBoard'
+import LogIn from './pages/LogIn/LogIn'
 
   export default function RoutesSet() {
 
     return (
       <Routes>
         <Route path="/*" element={<LogIn />} />
-        <Route path="/FormationBoard" element={<FormationBoard />} />
       </Routes>
     )
   }
