+++ src/components/Board/Board.jsx
@@ -9,11 +9,10 @@ import AddModal from '../Modal/AddModal'
 import Register from '../Register/Register'
 import axios from 'axios';
 
-export default function Board() {
+export default function Board(props) {
   const nodeRef = useRef(null)
   const [readyToModal, setReadyToModal] = useState(false)
   const [addModal, setAddModal] = useState(false)
-  const navigate = useNavigate()
 
   const onStart = e => {
     e.target.style.opacity = '0.5'
@@ -31,21 +30,9 @@ export default function Board() {
       setAddModal(true)
     }
   }
-  const onClickLogOut = e => {
-    axios
-      .post('http://localhost:3001/logOut')
-      .then(res => {
-        if(res.data.resultMsg == 'logOut_success'){
-          alert('로그아웃 되었습니다.')
-          navigate('/logIn')
-        }
-      })
-      }
-
 
   return (
     <>
-    <button onClick={onClickLogOut}>로그아웃</button>
       <div className={styles.wrapper}>
         <div className={styles.dragZone}>
           <ul className={styles.district_01}>
