--- src/components/Main/Main.jsx
@@ -1,8 +1,0 @@
-export default function Main(props) {
-  const id = props.id
-  return (
-    <>
-      <p>{id}님 가입하셨네염??</p>
-    </>
-  )
-}
