--- src/pages/FormationBoard.jsx
@@ -1,10 +1,0 @@
-import { Login } from "@mui/icons-material";
-import Board from '../components/Board/Board'
-
-export default function FormationBoard(){
-    return(
-        <>
-        <Board />
-        </>
-    )
-    }
