+++ src/pages/FormationBoard/FormationBoard.jsx
@@ -1,0 +1,11 @@
+
+export default function FormationBoard(){
+
+
+
+    return(
+        <>
+
+        </>
+    )
+    }
